Create Task4 mbot linnear calibration
This sequence has an initial light show that is mixed with a simple musical tune. then, there is a pause as the lights go red. The sequence which before the bot starts to move is inspired by the starts fo formula 1's races. the red light will shine for a short time and then, through random number generation  the yellow lights can change to green (go) at any time, keeping everyone on their toes. I did have some trouble with the *if...else* code but I will sort that out and would appretiate suggestions. Also, the reason I have the left wheel turn at 10% more power is because my bot turned heavily to the right, and in a distance of two meters forward, it also went off to the right side by at least a foot!