Create Task4 mbot angular distance calibration
This code is very basic and just consists of brainstorming because I have only tried forward motion once with the bot and never angular motion. I guess this will be very much like a tank and it can turn on a dime, or an area equal to a square with sides equal to the longest distance on the bot's proportions. I will try this all out on the bot on Friday.